feat: add confirmation dialog for prize deletion in PrizeModal to prevent accidental deletions

+++ base-commander/src/pages/rewards/prizes/PrizeModal.jsx
@@ -83,15 +83,18 @@ class PrizeModal extends Component {
   }
 
   onDelete = () => {
-    this.props.deletePrize( this.props.prize.prize_id )
-    .then( () => {
-      this.props.toggle();
-      this.setState({ saving: false, updates: {} }) 
-    })
-    .catch( e => {
-      this.setState({ saving: false } );
-      toast.error( e.message )
-    });
+    const conf = window.confirm( 'Are you sure you want to delete this prize? This action cannot be undone.' );
+    if ( conf ) {
+      this.props.deletePrize( this.props.prize.prize_id )
+      .then( () => {
+        this.props.toggle();
+        this.setState({ saving: false, updates: {} }) 
+      })
+      .catch( e => {
+        this.setState({ saving: false } );
+        toast.error( e.message )
+      });
+    }
   }
 
   render() {
